adding in service tests + changing Long to int

+++ src/main/java/org/mthree/dao/mappers/ItemMapper.java
@@ -10,8 +10,8 @@ public class ItemMapper implements RowMapper<Item> {
     @Override
     public Item mapRow(ResultSet rs, int rowNum) throws SQLException {
         Item item = new Item();
-        item.setId(rs.getLong("id"));
-        item.setUserId(rs.getLong("user_id"));
+        item.setId(rs.getInt("id"));
+        item.setUserId(rs.getInt("user_id"));
         item.setPlaidAccessToken(rs.getString("plaid_access_token"));
         item.setPlaidItemId(rs.getString("plaid_item_id"));
         item.setCreatedAt(rs.getTimestamp("created_at").toLocalDateTime());
